requirements.txt
Cython==3.0.10
unicorn-binance-websocket-api==1.46.2
unicorn-binance-rest-api==1.10.0
unicorn-binance-local-depth-cache==0.7.3
unicorn-fy==0.12.2
ccxt==4.3.2
pandas==2.2.2
hjson==3.1.0
SQLAlchemy==1.4.46
numpy==1.26.4
jsons==1.6.3
pybit==5.7.0
pyyaml==6.0.1
GitPython==3.1.43
pid==3.0.4
kneed==0.7.0
ckwrap==1.2.0
scikit-learn==1.4.2
influxdb-client[ciso]==1.30.0
requests==2.31.0
antropy==0.1.6
websocket-client==1.7.0
rel==0.4.9.9
ta==0.11.0
scipy==1.13.0
yarl==1.9.4
fastenum==1.1.1
pynput==1.7.6
# polars==0.19.0
sortedcontainers==2.4.0
numba==0.59.1
ujson==5.9.0
redislite==6.2.912183; platform_system != "Windows"
redis==5.0.3
setproctitle==1.3.3
# Streamsaver
pyarrow==15.0.2
# Dev
flake8==7.0.0
# pytest-bdd==5.0.0
bbd-table-parser==0.0.1
# Rest server
flask==2.1.2
Flask-Cors==3.0.10
# Backtesting
mplfinance==0.12.9b1
plotly==5.10.0
kaleido==0.2.1
prettytable==3.4.1
# UI
psutil==5.9.8
# Dashboard
dash_daq==0.5.0
dash==2.9.1
# kite
kiteconnect==4.2.0
pyotp==2.9.0
Werkzeug==2.2.2
pycryptodome==3.20.0
cryptofeed==2.4.0